+++ src/components/map-legend/mobile/style.module.scss
