+++ src/components/map/component.js
@@ -2,6 +2,7 @@ import React, { PureComponent } from 'react';
 import MapGL, { NavigationControl } from 'react-map-gl';
 import PropTypes from 'prop-types';
 import MediaQuery from 'react-responsive';
+import LegendMobileControl from 'components/map-legend/mobile';
 import classnames from 'classnames';
 import { breakpoints } from 'utils/responsive';
 import BasemapSelector from 'components/basemap-selector';
@@ -81,8 +82,13 @@ class Map extends PureComponent {
         <div className={classnames(styles.legend,
           { [styles.expanded]: !isCollapse })}
         >
-          <Legend />
-          <BasemapSelector />
+          <MediaQuery maxWidth={breakpoints.md - 1}>
+            <LegendMobileControl />
+          </MediaQuery>
+          <div className={styles.tooltip}>
+            <Legend />
+            <BasemapSelector />
+          </div>
         </div>
       </MapGL>
     );
