+++ src/components/map/style.module.scss
@@ -28,6 +28,10 @@ $space-left-map-mobile: 20px;
       right: 25px;
     }
   }
+  .tooltip {
+    display: flex;
+    flex-direction: column;
+  }
 }
 
 // Mapbox override
